+++ doc/changelog.rst
@@ -3,6 +3,7 @@
 Release Notes
 =============
 
+- :bug:`1259` The organiser dashboard included deleted proposals in the count on the event overview.
 - :bug:`-` Some tables were not responsive on mobile devices, particularly in the organiser area.
 - :feature:`1184` Review forms can now be sent with ctrl+enter for a smoother review workflow.
 - :bug:`1253` Review scores were not copied when choosing a previous event to copy to a new event.